+++ src/app/features/pages/shuttle/shuttle.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import Toast from 'awesome-toast-component';
 import { NgxSpinnerService } from 'ngx-spinner';
+import { RouteService } from 'src/app/core/services/route.service';
 import { ShuttleService } from 'src/app/core/services/shuttle.service';
 
 @Component({
@@ -13,6 +15,7 @@ export class ShuttleComponent implements OnInit {
   constructor(
     private router: Router,
     private shuttleService: ShuttleService,
+    private routeService: RouteService,
     private spinner: NgxSpinnerService
   ) { }
 
@@ -31,16 +34,27 @@ export class ShuttleComponent implements OnInit {
   }
 
   async deleteShuttle(shuttle: any){
-    (await this.shuttleService.deleteShuttle(shuttle.id)).subscribe(()=>{
-      this.setRegisteredShuttles();
-    });
+    this.routeService.getRouteByShuttleId(shuttle.id).subscribe(async (res)=>{
+      if(res.length > 0){
+        new Toast("Error: Cannot delete shuttle as it is connected to a route. Please delete the route first!", {
+          position: 'top',
+          theme: 'light'
+        });
+      }
+      else{
+        (await this.shuttleService.deleteShuttle(shuttle.id)).subscribe(()=>{
+          this.setRegisteredShuttles();
+        });
+      }
+    })
+    
     
   }
 
   async setRegisteredShuttles(){
     this.spinner.show();
 
-    (await this.shuttleService.getAllShuttles()).subscribe((res: any[])=>{
+    (this.shuttleService.getAllShuttles()).subscribe((res: any[])=>{
       res.forEach((doc: any, index: number) => {
         if(doc.shuttleImage != ""){
           this.shuttleService.getImageDownloadUrl(doc.shuttleImage).then((observable)=>{
@@ -51,21 +65,10 @@ export class ShuttleComponent implements OnInit {
         }
         else{
           res[index].downloadUrl = "no image";
-        }
-
-        // if(doc.shuttleImage != ""){
-        //   (await this.shuttleService.getImageDownloadUrl(doc.shuttleImage)).subscribe((value: any)=>{
-        //     let imagePath = JSON.parse(value);
-        //     res[index].downloadUrl = imagePath.image;
-        //   });
-        // }
-        // else{
-        //   res[index].downloadUrl = "no image";
-        // }
-        
+        }       
         
       })
-
+      console.log(res);
       this.listOfShuttles = res;
       this.spinner.hide();
     });
